(#111) Rename namespace to PositionNames that PositionName model can be used for CreatePositionName API
